added core folder and moved appropriate models there
