+++ BackEnd/profiles_service/routers/profiles.py
@@ -1,6 +1,6 @@
 from typing import Annotated, List
 from fastapi import APIRouter, Depends, Response
-from exceptions import NotFoundError, InvalidRelatedEntityError
+from core.exceptions import NotFoundError, InvalidRelatedEntityError
 from schemas.profiles import StartupOptional, Startup, ModerationFeedback, StartupResponse
 from services.profiles import ProfileStartupService
 from dependencies import get_startup_service, startup_create_dependency, startup_optional_create_dependency
